requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
